// frontend/src/types/CustomerItem.ts
export interface CustomerItem {
  id: string;
  customer_id: string;
  displayId: string;
  fullName: string;
  email: string;
  phone: string;
  gender: string;
  avatar: string;
  username: string;
  role: string;
  address: string;
  brandId: string;
  brandName: string;
  status: string;
  totalOrder: number;
  dateJoined: string;
  dateOfBirth: string;
}